==> walmart_M5_forecasting_lgbm/4_apply-factors-m5.ipynb <==
dirct = r'D:\python_code\M5 Competition\m5-forecasting-Dataset\\'

import numpy as np
import pandas as pd 

import os
for dirname, _, filenames in os.walk(dirct):
    for filename in filenames:
        print(os.path.join(dirname, filename))

pd.options.display.max_columns = 50
pd.options.display.max_rows = 50
import warnings
warnings.filterwarnings('ignore')

sub = pd.read_csv(dirct + 'partial_submission_18nov.csv')
factors = pd.read_csv(dirct + 'factor_computed.csv')
df_ids = pd.read_csv(dirct + 'sales_train_evaluation.csv', usecols=['id','dept_id','store_id']).drop_duplicates()

sub = sub.drop_duplicates()

sub

factors = df_ids.merge(factors, on=['dept_id','store_id']).drop(['dept_id','store_id'], axis=1)
sub = sub.merge(factors, on='id', how='inner')

sub

cols_f = ['F{}'.format(c) for c in range(1, 29)]
for c in cols_f:
    sub[c] = sub[c]*sub['factor']
    
sub = sub[['id']+cols_f]

sub

sub2 = sub.copy()
sub2["id"] = sub2["id"].str.replace("evaluation","validation")
sub = pd.concat([sub2, sub], axis=0, sort=False)
sub.columns = ['id'] + ['F' + str(c) for c in np.arange(1,29,1)]

sub.to_csv(dirct + 'submission_18nov.csv', index=False)

sub
